GitOperations/requirements.txt
pandas
requests
tqdm
beautifulsoup4
psycopg2
PyGithub
click